Almost ready to merge with main branch now, These are mostly updates to the docs, with some compiler fixes and general warning suppressions.
[SVN r26648]

+++ src/cregex.cpp
@@ -567,9 +567,9 @@ std::string RegEx::What(int i)const
 }
 
 #ifdef BOOST_HAS_LONG_LONG
-const std::size_t RegEx::npos = ~0ULL;
+const std::size_t RegEx::npos = static_cast<std::size_t>(~0ULL);
 #else
-const std::size_t RegEx::npos = ~0UL;
+const std::size_t RegEx::npos = static_cast<std::size_t>(~0UL);
 #endif
 
 } // namespace boost
